kafka/client: Move a kafka publishing util into kafka/client
